heuristic to choose base channel and delta type
This program offers a way to choose which channel to use as a reference and whether to use horizontal or vertical deltas.  Looks like the decision about the deltas is more important.  If the image was generated using a bayer filter, using the red or blue channels as a reference might reduce detail.