checking whether channels are correlated
Delta encoding works better or worse with inter channel deltas depending on whether channels are directly or inversely correlated.  To get the the optimal compression we need to use a metric to decide which deltas to use.  It is also appears that the recursive bitwise compression gets a better compression rate than huffman codes with the unary strings in some situations.