more information about different compression rates
Seems pretty clear that compressed unary strings produce more compression than huffman codes alone when the optimal prefix-free code rate deviates from the shannon limit, which it does frequently when delta encoding natural images.  What wasn't clear is that the extra compression approaches the shannon limit regardless if run bits or stop bits are being compressed.  The improvement seems to range from 0 (when the prefix-free code rate is the same as the shannon rate) up to 25%, when the difference between the compression rates is large.