debugging
Implementing Huffman encoding because 1) the results from Deflater using Huffman-only are larger than the Huffman cost function (except for binary images, when it's smaller) and 2) to see if Huffman codes can be compressed with the bitwise transform, which I always doubted but never checked.  Don't want to rely on testing Deflater output to draw conclusions because of the discrepancies noted above.  If we can successfully implement an interface for Huffman coding that also works with unary strings, we should also end up with a framework where any variable length code is easily supported,.