implemented huffman coding
Implemented huffman coding so we don't have to use results from gzip in hffman-only mode as a proxy for simple huffman coding.  Looks pretty obvious from the zero ratios that they wont compress further with the bitwise transform.  Have a neat compression interface now to use any random variable length code, although it lacks a decompressing function.  The reason I know it works is it produces the exact same results as the original parsing functions for the unary strings.