modified paeth filter
Modified the paeth filter to check the forward diagonal pixel as well as the backward.  It produces a small improvement on smooth images, but a poorer result for more random values, which was unexpected.  Don't exactly understand why using more information in the predictive process would produce worse results with random data--maybe feedback loop based on mistakes would help.