zipping compressed strings
Looks like zipping the packed or packed/compressed strings from the unsegmented data produces the best compression.  The segmentation is intriguing because it does improve the compression, but seems to be exactly balanced by the increased overhead.  Useful for breaking the data up into packets, but not improving overall compression unless there's a way to decrease the overhead.