Fixed a bug.
Fixed a bug that made the delta 8 function appear to do very badly.  Seeing a huge disparity in the amount of compression of the separated maps versus the combined map, which is hard to understand.  If it isn't a bug, it seems to correlate with how large the seed segment is.