added segmentation parameter
Revisiting segmentation to see if there's some combination of delta type and zipping that significantly improves the compression rate of the unary strings.   So far, it seems that there is only a slight improvement, but that's still a useful feature for networks that the data can be packetized and actually reduce the over all amount of data.  If there was actual significant compression as well, it would have to of interest to someone in telecommunications.