added opportunistic zipping
Huffman coding compressed unary strings and then sending the result or not depending on whether it improves the compression, which it does not always do.  Although the improvement is not that significant it is another proof that there is recursive compression.  It also seems to be the key to allowing segmentation to significantly improve the compression rate, at least in the case of delta type 7.