figured out how to reduce header size
With a large number of segments and small segment sizes it helps to group segment lengths in a block and zip them.  Also it appears there might be some way to improve the merge process since even without  headers, large numbers of segments are producing slightly less compression.  It might be that we want to do something besides just merge two segments together whenever they are similar, and spend more time searching for segments with different types.  In other word if most of the segments are type 0, go back and look for as many segments of type 1 as possible.