cleaned up code
Removed code that merges segments using the smallest possible bit lengths.  Still possible that it would improve compression rate, but probably not much and it's a lot work and makes the code complicated.  Probably easier to improve compression by grouping segment lengths together when saving the data and also working with how deltas are produced.