comparing huffman vs recursive unary string coding
It looks like the advantages of unary strings are with small numbers of elements when the smaller overhead (rank table vs. frequency table or dictionary) comes into play, or with a more skewed distribution when the recursion of stop bits comes into play.  Recursive compression of run bits also helps improve compression rates for uniform distributions, making it more practical for lossless compression.