more fiddling with segmentation
Tried joining segments together to eliminate trailing bits when segments are stored separately.  Improves the compression rate by tens of thousandths, as does zipping the segment length and info in blocks instead of sending them separately unzipped.  The compression rate for segmented data is almost exactly the same as unsegmented data when the overhead is carefully managed, within tens of thousandths.  Useful that the total amount of data sent over a network would be the same if the image is broken up into separate packets.