implemented bottom-up huffman encoder
Decided to stop using gzip (in the form of Java Deflater/Inflater) so I can be sure what extra overhead is involved without having to crawl through the java native methods somehow or png c code.  Also, the constraint that input only be 8 bits long is unnecessary, and makes it impossible to process deltas from unquantized 24 bit images when the value range is greater than 256.