Finished instrumenting string compression.
Some puzzling results compressing stop bits--it looks like the bitwise transform can produce zero compression after one pass, and then go negative.  Also, sometimes the string with be further reduced by the inverse transform when we expect it to expand.  Both of these properties offer an opportunity for a modest improvement in compression rates.  The compressed run bits seem to behave as expected.