added histogram
Producing histograms of the packed unary strings, compressed packed strings, and segmented compressed packed strings.  As might be expected, the histograms become progressively more uniform.  The segmented string has an outsize number of 0 values that can be explained by empty bits in the concatenation. We expect dropping the empty bits will bring the number of 0 value closer to the mean.