Finished implementing splicing.
Splicing bits from uncompressed segments to compressed segments reduces the overall bit length by a small amount, but makes the segments irregular which makes further processing a little trickier.  Might offer a way to reduce overhead, if the overhead for an uncompressed segment is more than its bit length plus any increase in the bit length of the compressed segment.