anomolous stop bit strings
It looks like a fully compressed stop bit string can be further compressed by the run bit transform.  We considered that compressOneStrings and decompressZeroStrings did exactly the same thing, but apparently not.  Should allow for a modest improvement in the compression rate, even if we don't understand how it happens.