finished adding new methods
Still need to remove some of the lower level processing in SegmentMapper().  Ready to do the extra iteration on stop bit strings in the opposite direction to get a modest improvement in the compression rate, although it will limit the number of iterations in either direction to 8.   That will only be an issue when applying the transform to larger strings, such as the data from an entire channel.  Still appears to be something that doesn't happen with run bit strings but don't see how that helps save space in the data byte.