Implemented anomalous string handling
Still need a way to poke an already compressed segment to find out if its anomalous to make it usable in the image compression.  Don't want to incorporate the information in the data byte because we would have to reduce the possible number of iterations by half to add a special bit, and at low color resolutions we are already coming close to maxing out.