debugging anomalous string handling
Running into an off by one error on run bit compression/decompression.  It either works with all the values from a channel, or it fails with all the values in a channel, which might offer a wedge into figuring out what's going on.  Trailing bits from recursion would add bits, not subtract them, and the compressed/decompressed segment is one bit shorter than it should be, so we don't think that's the issue.