more anomalous string handling
In the course of trying to apply both run and stop bit transforms to the same segment, we have apparently discovered that simply decompressing and recompressing zero strings often reduces the bit length.  Obviously the underlying primitive routines are not doing exactly what we thought and we'll have to double check it.  It complicates  handling anomalous segments but still feels like a lucky accident.