added option to process channels w/o bit transforms
It looks like the unary strings from an  entire channel sometimes or usually compress more with deflate without applying the run bit or stop bit transform.  Probably the more homogenized the string string is the more effective the bit-wise transforms.  It might be that the main utility for the bit transforms is in the segmentation, when they can be applied more adaptively.