anomalous string handling working
Managed to compress strings by first compressing stop bits and then compressing one bits, which conclusively demonstrates that stop bit decompression is not the equivalent of run bit compression.  Keeping the anomalous information in the data byte, after failing to find a way to figure it out by poking the raw data somehow.  Trading off number of possible iterations to keep information about anomaly, which would only reduce the compression rate when  the color resolution is low enough to produce sufficiently sparse matrices that the number of stop bit transforms might be that large.