more support for java-only huffman coding
We can do all the huffman coding we want now without using deflate, and we are also not constrained to byte input.  Deflate() is a black box unless we navigate the native methods and translate them into java, and we can get similar results without it and we know how we get them.  It seems like the image compression has very little to do with deflate lzw facilities, which are mostly relevant to text compression.  There might be a place for pattern recognition in the delta encoding but lzw-like methods do not seem to apply in an obvious way.