fixed a bug
Doing histogram of unsigned bytes instead of signed bytes.  This fixes an out of bounds error when referencing the rank table, and might be improving the compression rate.  The signed byte histogram would probably work with some extra processing, but starting off with unsigned bytes seems like the simplest approach.